+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/groq
 
+## 4.0.0-beta.20
+
+### Patch Changes
+
+- Updated dependencies [176466a]
+  - @ai-sdk/provider@4.0.0-beta.10
+  - @ai-sdk/provider-utils@5.0.0-beta.16
+
+## 4.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [e311194]
+  - @ai-sdk/provider@4.0.0-beta.9
+  - @ai-sdk/provider-utils@5.0.0-beta.15
+
 ## 4.0.0-beta.18
 
 ### Patch Changes